Rename the rule and update one test
